MCObject::load() shouldn't be called for MCStack instances
When loading a stack, the type byte needs to be passed to
`MCStack::load()` in order to allow the correct stack loading mode to
be selected.  All other objects can be fully loaded without the type
byte.

Previously, only a 3-argument variant of `load()` was defined in
`MCStack`, which meant that it was possible to (incorrectly) call the
base class (2-argument) implementation of `MCObject::load()` by
upcasting an `MCStack` pointer.  This was flagged up by the emscripten
C++ compiler with a "overloaded-virtual" warning.

This patch explicitly overrides `MCObject::load()` in `MCStack` with
an implementation that indicates a programming error.