[[ Bug 15612 ]] Reinstate MCtodestroy stack list.
When reworking object deletion to be more prompty, the pending destroyed stack list was replaced by the same
object deletion process. However, this was incorrect. MCtodestroy is a list of stacks which are pending
deletion due to the 'destroyStack' requirement being satisfied. These are always deleted at the next main
loop iteration, rather than immediately. (This ensures destroyStack works even if handlers in the stack are
currently executing).