[emscripten] LiveCodeAsync.resume(): Allow calling without pausing first
Sometimes the browser can fire a callback function either immediately
or at some point later, and it's not clear which (or behaviour might
vary between browsers).  Make sure that bad things don't happen if
LiveCodeAsync.resume() is called when the engine isn't actually
paused.

This has the advantage of slightly simplifying the logic for
LiveCodeAsync.delay().