[[ Script ]] Allow per-module native code libraries.
The user of libscript can now set a callback to be used to resolve the path to
native code libraries on a per-module basis. This is used when binding foreign
handlers to libraries with no path, the callback is invoked with the module
context so that a mapping to an absolute path for the code module for the
current platform can be computed.