[[ Business ]] Add "additional manifest" command to installer syntax
This (in theory -- currently untested!) adds the contents of the
given file as they were written at the top level of the original
manifest file.

Because they are included at top-level and no merging takes place,
the main use of this is to add new components that can be referenced
via "include" statements in the main manifest.