MCNotify: Gate API use with FEATURE_NOTIFY macro.
On some platforms, such as emscripten, no threading API is available
(and threads aren't needed anyway).  In that case, there's also no
need for the `MCNotify` API.

This patch gates use of the `MCNotify` API using the `FEATURE_NOTIFY`
macro, and ensures that the API implementation won't compile on
platforms that don't define `FEATURE_NOTIFY`.