[[ Bug 15675 ]] Crash when 'DataTree' is in Plugins
The IDE crashes shortly after startup if DataTree is in the plugins. This appears to be
because at some point DataTree uses a stack with 'destroyStack' true, which causes
the current tooltip to be set to NULL. However, the 'settip()' method used expects a
valid stringref and not NULL. This call has been replaced with 'cleartip()' which fixes
the problem.