[[ Hooks ]] Generalize Android native control creation code.
The methods such as 'createBrowserControl' in Engine have been refactored to use a single
'createNativeControl' method in the NativeControlModule. This method takes the fully-
qualified name of the class that should be instantiated to create the native control. The
class must derive from NativeControl and have a constructor taking a NativeControlModule
as an argument.