[[ Hooks ]] Add 'signature' field to MCNativeControlActionInfo and use that to dispatch.
The MCNativeControlActionDo method relies on finding an appropriate MCNativeControlActionInfo
struct. The latter describe specific actions by mapping a tag (MCNativeControlAction) to a
method pointer.

The code has been updated to add an explicit signature field to the Info struct which is
use in MCNativeControlDo to dispatch properly. As most actions share similar signatures
this makes much more sense.