[emscripten] Download the standalone capsule before running engine.
Emscripten provides a "run dependency" mechanism for preventing the
program from being started until a set of dependencies are filled.
This is used by Emscripten for downloading the initial memory image
and any filesystem files provided via emcc's `--preload-file`
argument.

This patch makes the emscripten-js standalone engine use the same
dependency mechanism for downloading the standalone capsule (which is
expected to be a zip file containing the boot filesystem image for the
engine).  `em-preamble.js` is included at the start of the compiled
JavaScript file and is run before anything else.

The new code defines several new entries in the `Module` object:

* `livecodeStandalone`: the filename for the standalone zip file

* `livecodeStandalonePrefixURL`: prepended to `livecodeStandalone` to
  create a URL

* `livecodeStandaloneRequest`: The `XMLHttpRequest` instance for
  downloading the standalone file.