[emscripten] When not installed, use original standalone template
The build process copies the standalone template directory into the
top level `emscripten-bin` directory.  However, it doesn't check for
changes to the original directory and update the copy.

This patch makes the standalone builder use the original version of
the template in `engine/rsrc` rather than the copy, when running from
the git repository.