[[ Bug 15799 ]] Array containing empty array encoded in 7.0 won't load into 6.7.
The 7.x engines are currently (in some cases) encoding empty arrays as an array with 0 elements
when encoding the array in 6.7 format. Unfortunately, this is not quite compatible with the
previous method 6.7 uses to decode - where it assumes that empty arrays will not appear at all
and be empty strings instead (which is consistent with the 6.7 execution model).

This has been fixed by making sure the 6.7 engine looks for the terminating byte of the array
if it has 0 elements.