[emscripten] Run startup script in temporary stack.
Currently, widgets are instantiated on stack load.  This means that
all extensions required by a stack must be loaded with `load
extension` *before* attempting to load the stack itself.

This means that the emscripten standalone `__startup` script needs to
be run on a temporary `MCStack` instance *before* attempting to load
the standalone's main stack.