[emscripten] Improved JavaScriptification script in Python.
Replace the existing `emscripten-javascriptify` with a new one written
in Python.  This is useful for commercial builds, because it makes it
easier to pass in arbitrary numbers of both preamble JS and library JS
files to the compiler.

I considered doing this in Bash but decided it would be faster and
less error-prone to just port it to Python.