[[ Bug 13537 ]] MCLocaltoken::parse: do not check for global var out of scope
With Strict Compilation turned on, local variable creation should only check for
the existence of global variables declared in that handler and in this script only.

Checking whether the global variable is made by calling either `sp.gethlist->findvar` or `sp.gethandler->findvar`
The removed code was over-zealously checking for the out-of-scope global variables: they have to be declared global in
the handler or the object script to be recognised as such.