[Bug 16092][emscripten] Set the canvas to use the "default" cursor
By default, the browser sets a contentEditable element to have the
"text insertion" cursor.  Because the canvas is set to be
contentEditable in order to receive text input events, it's necessary
to explicitly set the canvas back to using the default cursor.

A subsequent patch should connect up the cursor control syntax in
LiveCode to control the CSS cursor.