[[ Mac64 ]] Replace the removed GetCaretTime() function
There isn't a direct replacement, so an undocumented preferences
key needs to be queried instead (NSTextInsertionPointBlinkPeriod).
This key is widely discussed on the internet so seems to be fairly
reliable.

Just in case it isn't, a fallback value is provided that uses
the default blink rate for my OSX system.