[[ RawClipboard ]] Restore the buggy clipboardData behaviour
Due to a bug in the old code, images were never chosen when text
was also available. This bug has existed for long enough to become
ossified as a feature so this patch restores that behaviour.

The code that implements the behaviour originally intended is
still present but disabled, just in case we ever want to switch
to using it.