[Tests] Tweak files test to avoid spurious TOCTTOU failures.
Previously, `the files` was being tested using the contents of the
system temporary directory.  This test was occasionally failing due to
time of check to time of use errors -- files that were present at the
time `the files` was evaluated had been removed by other system
processes by the time `there is a file` was evaluated.

Instead, just use whatever the current working directory of the test
runner is.