[[ Bug 7414 ]] Focusing a listBehavior field doesn't set the active field.
It would appear that when listBehavior was added to the field, the code
which sets the 'active field' to the focused field ended up being in the
wrong place.

The result was that focusing a list field by any other means than clicking
on it would result in the focusedObject being the field, but the
selectedField not being the same.