[[ Bug 16257 ]] Windows do not unfocus correctly.
A problem with window focus switching was introduced with a fix to
Bug 15208. This fix essentially stopped windows receiving 'unfocus'
events.

Reversion of that patch uncovered a typo in 'windowDidResignKey'
which would cause ProcessDidBecomeKey to be sent instead of
ProcessDidResignKey in some instances.

It is likely that this error would occasionally cause other focusing
related issues.