[[ Bug 16270 ]] Crash when mouse scrolling a composed widget.
When mouse scrolling over a composed widget which recreates its children,
it was possible for MCwidgeteventmanager to still hold references to an
unplaced child in mouse focus, or mouse grab.

This patch cleans up synchronization of state after unplace / place,
ensuring that the event manager holds the correct focused, or grabbed
control after placement or unplacement. At the moment this is done
immediately, however it might be better that it work via a deferral
mechanism at some point to improve efficiency.

Additionally, it also ensures that OnOpen and OnClose are called when
widgets are placed or unplaced into open widgets and tightens up the
treatment of mouse clicks. Ensuring that the original button used to
start a click is the one which is presented when all buttons are
finally released.