buildbot: Tweak rules for generating checksums for uploaded files
The new rules for generating and uploading a SHA-1 checksum of
uploaded installers were causing build failures due to (1) the lack of
a `sha1sum` program on Mac OS X, and (2) incorrect `rsync`
command-line syntax (you can't use both `--files-from` and filenames
as arguments).

This patch modifies the logic so that the checksum list isn't
generated when running the build on Mac, and generates the checksum
file and the list of files to upload in the same rule -- ensuring that
the list of files to upload includes the checksum file.

(cherry picked from commit 9036b0db42934895534e436030679058104b89ff)