[[ Bug 15097 ]] Make 'the effective rect of stack' more reliable on Linux
Before the port to GDK, the Linux engine used its own heuristic for computing
the frame rect of a window - this was not entirely reliable.

The appropriate method has now been updated to use gdk_window_get_frame_extents
which uses a couple of methods to try and ensure accuracy. Indeed, as long as
the window manager is compliant with EWMH (and isn't broken) it should be
correct.