[[ Mac64 ]] Update the externals API to be 64-bit clean
The window IDs were being passed as 32-bit values but this isn't
guaranteed to work on all platforms (in particular, it would break
for a 64-bit Windows port).

This is technically an ABI-incompatible change for 64-bit Linux
*but* the only affected external shipped by us is revbrowser
(which is new in 8 and we're still in DP) and there are no other
Linux externals that I know of that use this particular call.