Add utility for updating submodule pointers.
Add a new tool that loops over all "develop" or branches in a remote
repository, and update their submodule pointers.

It won't do anything unless the submodules have been configured with
appropriate "branch" keys in the `.gitmodules` file for that
particular branch.  At the moment, only the "develop" branch has the
`.gitmodules` file configured appropriately.

N.b. you *must* run this on a *completely clean* repository because it
*nukes everything* (it'll warn you if there's any chance that you'll
lose work).