libfoundation: Fix finding single-character needle in string.
In the single-character needle case, `MCUnicodeFirstIndexOf()` was
dereferencing 1-element `char_t` or `unichar_t` array as a
`codepoint_t`, resulting in a 2- or 3-byte buffer overrun.

In addition, the single character codepath is only valid if the the
needle is (1) a native ASCII character or (2) a BMP Unicode character.
This wasn't being tested correctly.