[[ Tests ]] Add extension testing capabilities
This patch adds extension testing to our continuous integration, by adding a further check target in the
extensions directory. At some point the tests/Makefile will need tweaking to make sure that only
<extension>/tests/ directories are searched for suitable test handlers.

Also TestLoadExtension has been added to the test library to facilitate loading of extensions in the
TestSetup handler. At some point it might make sense for extension tests to automatically load themselves
but for now it seems fine to have them do it explicitly.

Note this does no dependency analysis, so it is currently not possible to test extensions that rely on modules
other than those built in to the engine.