[Bug 16021] segmented: Correctly compute icon bounding box
The bounding box for the icons in the segmented widget were being
computed incorrectly in the case that the segment width was greater
than the segment height.  This caused weird things to happen as the
widget was resized.

This patch changes the code to first find the largest icon size that
will fit, and then compute the appropriate icon bounding box
accordingly.