[[ Bug 16333 ]] Rework the way chunks are read from streams
New IO_handle created: MCBufferedFileHandle

This IO_handle can put back chars, which comes handy (if not necessary) when reading characters from Unicode
stream: we need to read one codeunit over the char boundary to ensure that we have read a whole char, and
the former way was not preserving this codeunit after a `read from file <x> for <X> chars`.

`MCFilesExecPerformReadChunk has been updated to read only what it is asked for, and will put back what it reads
over the char boundary.

Every OpenFile, OpenDevice and OpenFd system function have been updated to open an MCBufferedFileHandle instead
of a normal MCStdioFileHandle