[Bug 16021] segmented: Fit non-square icons correctly.
The fix in commit 205e7fc70c2c was correct for square icons, but would
result in pessimistic fitting for non-square icons.  This patch
changes the logic to compute the minimum necessary padding within the
segment, and then the maximum icon bounding rectangle that provides
that padding.  This gives the same result as before for square icons,
but allows non-square icons to be as large as possible.