[[ Builder ]] Add docs data to sqlite database on installer build
There is no easy way of creating a docs api in the IDE using the
JSON data, so we include an sqlite database with each entry's docs
data array arrayEncoded for easy retrieval.

The database includes a table of libraries (currently livecode script
and livecode builder) and a table for each library containing its
docs data.