[Bug 15819] Remove shell special chars from Linux install path
Having space and parentheses in the default Linux install path makes
things really awkward when writing shell scripts or Makefiles that
need to refer to paths within the LiveCode installation directory.
It's hard to get the escaping right and it's hard to explain to
people.

This patch changes the default installation path from e.g.

    /opt/livecode/livecodecommunity-8.0.0-dp-9 (x86-64)/

to

    /opt/livecode/livecodecommunity-8.0.0-dp-9.x86_64/

This has the added benefit that the LiveCode executable is named the
same as the directory it lives in.