Ensure use clause doesn't clobber use definition
This patch resolves the issue where if a use definition was specified
prior to any other definition the parser would think it is a use
clause and throw an unknown use error and then see the on and think
it is an `on x call y` definition. A side effect of this patch is the
parser will no longer require use clauses to be declared before anything
else. Call is still required before anything but use.