libbrowser: Don't use aliased and never-defined structure types.
Previously, `MCBrowserListRef` was defined as a pointer to a `struct
__MCBrowserList` that was never defined anywhere.  Instead, every use
of `MCBrowserListRef` was aliased to a complete `class MCBrowserList`
type.  There were several other types used in a similar way by
libbrowser.  This highly confused static analysis tools!

This patch defines each `MCBrowser*Ref` type to be a pointer to the
corresponding class type, and removes as many unnecessary aliasing
casts as possible from libbrowser method functions.