Resolve regression to use clause
This commit resolves a regression where a use clause is followed by an
on .. call ... clause the parser will look ahead to the next line
and treat the on as part of the use definition. The option to ensure
the matched token is in the same row as the parsers current position
has been added to all the ParserWillMatch... functions. There are very
likely other functions in the parser that could make use of the extra
option.