[[ Tests ]] Fix log encoding issues.
This patch ensures that tests can output UTF-8 into the log file
and that line endings are more appropriate on Mac.

Code that writes the log from _testlib now encodes the text
explicitly as UTF-8 as 'stdout' is considered to be native encoded.
The testrunner then decodes the output from the subprocesses
appropriately.

The log file is now written with LF line endings on Mac and Linux,
with Windows remaining as CRLF.