[[ Docs Data ]] Add code to docs parser stack to create or update a docs database
There is no easy way of creating a docs api in the IDE using the
JSON data, so we include an sqlite database with each entry's docs
data array arrayEncoded for easy retrieval.

The database includes a table of libraries (currently livecode script
and livecode builder) and a table for each library containing its
docs data.

The reason for adding it here is that it will be accessible both while building
the installers and running the IDE.