[[ Benchmarks ]] Add support for variants of a single benchmark.
A single benchmark command may now issue multiple StartTiming and
StopTiming calls, passing an optional parameter 'variant'.

The log output has been changed to be lines of the form:

<BenchmarkName> tab <Variant> tab <TimeInMS>

If no 'variant' parameter is given to StartTiming, it is presumed
to be empty.

Additional a 'BenchmarkLoadNativeTextFile' command has been added
which loads the given file, searched for relative to the benchmark
script file, as a native encoded text file.