[CID 120051] Fix integer overflow coverity issue.
After the rework of native string handling, the implementation of
the Hash algorithm was based upon the word size of the architecture
rather than the memory model.

The memory model macros (__SMALL__, __MEDIUM__, __LARGE__) have been
corrected to be either __SMALL__ for 32-bit, or __MEDIUM__ for 64-bit
and __MNativeStr_Hash now keys the 32-bit or 64-bit hashing algorithm
on the __LARGE__ memory model where hash_t is 64-bit rather than 32-bit.