Mapped data type support
A long time ago some work was done on mapping generalised data types to
native data types e.g.
 - string
    -> objc-string on iOS and Mac
    -> c-string on Windows and Linux
    -> jstring on Android

These types use the `use <language> on <platform>, ...` definitions to
map the generalised type to the appropriate object type for the language
which enables the use of a single handler definition to apply to
multiple languages. All the work was done apart from the error checking
on the types being declared was throwing an error on the generalised
types.