[Bug 16076][emscripten] Generate Emterpreter whitelist more cleverly
Apply two sets of regular expressions to the symbols in the engine:

- Only symbols that match a whitelist expression are compiled with
  Emterpreter

- No symbol that matches a blacklist expression is compiled with
  Emterpreter

The set of regular expressions provided in this patch seems to cover
most possible ways that the engine could end up waiting, but it's not
guaranteed to be exhausted.  Just compiling everything in the core
engine to Emtepreter bytecode made the engine very large and very
slow, so a more selective approach was required.