[[ Bug 15429 ]] Fix compilation of nested invokes in inout contexts
Code which passes nested invokes to an inout argument (whether it be
for an LCB handler, or to a foreign handler via syntax) now compiles
correctly.

The code generation phase was freeing expression register attachments
too soon, meaning that the assignment phase of code generation was
unable to find the appropriate register to use.

Additionally, a consistency check has been added to ensure that when
a handler has finished being compiled all register attachments have
been detached.