[Perf] libfoundation: Better __MCNativeChar_CheckedFold() accuracy
This patch modifies `__MCNativeChar_CheckedFold()` to return a more
accurate result.  It now looks up the input char in a constant bit
array in order to determine whether it needs to be folded.  The
constant bit array is only 32 bytes, which means it fits into half a
cache line.  Because `__MCNativeChar_CheckedFold()` now only calls
back to `__MCNativeChar_Fold()` if necessary, searching strings for
this change should provide a speedup for case-sensitive string
operations on inputs that contain a large amount of digits or
punctuation.