[Bug 16696] Ensure that the correct release notes are installed
Change the package manifest to copy the release notes directly from
the top level of the livecode repository.  This is where the
freshly-built release notes are placed by the release notes builder.

Because the file that's copied into the installer has the version
number in its filename, this should also ensure that the installer
always contains the correct version of the release notes.