Always used themed back colours for controls
Before 8.0, all controls had a background colour of white if it
wasn't explicitly set (unless the theme was Motif). In other words,
the back colour wasn't inherited.

Some of the theming changes made in 8.0 caused the property to
become inherited, which was particularly noticeable for fields.
This patch changes the code to always used the appropriate themed
colour for the background (we don't use white as some themes, like
some versions of Ubuntu, use a light grey for field backgrounds).