[Bug 15970] New logic for choosing stack file saving format
If no stack format is requested, try to use the latest one available.

If a specific stack file format is requested, try to save in that
format, and produce an error if doing so would cause data loss.  At
the moment, this means that saving a stack produces an error if the
stack contains widgets and 8.0 format wasn't requested.

If 8.0 stack format is requested, but no widgets are present, use 7.0
format.

Set the `stackfileversion` to "8.0" by default, so that `save stack`
preserves widgets.