[[ Bug 16351 ]] Make MCtargetptr an MCObjectPtr
The initial work to ensure that 'the target' keeps track of the
part id as well as the object pointer. This will prevent 'the long
id of the target' from always resolving to the instance on the
current card when the target is a shared group.

This does not yet go all the way to fixing bug 10822 however, as
the part id is not yet passed as a parameter into intenal MCObject
functions such as message and dispatch, or when retrieving custom
properties (as in the example for bug 10822)

It does do the groundwork for fixing 10822 however, and it fixes the
issue in the specific case when the dispatch / send / call syntax is
used in script.