[Bug 16739] Ensure widgets in substacks are saved.
Although there was an `MCStack::substackhaswidgets()` method defined,
`MCObject::haswidgets()` wasn't calling it.

When saving stack files with widgets only in substacks (not in the
main stack), the save logic thought that no widgets were present and
downgraded the file format to 7.0, causing widgets in substacks to be
discarded.

This patch makes `MCObject::haswidgets()` virtual and overrides it in
`MCStack` with an implementation that calls `substackhaswidgets()`.