[[ Nav Bar ]] Allow navData of nav bar to be empty
The various rect calculations have been rejigged to avoid using
the navData list to store the base paths for the add, delete and
reorder actions - these are stored in a separate array and then
translated as necessary to calculate the path for each row (in
edit mode)

When the navData list is empty, only the add item icon path is
calculated.

For normal (non-edit) mode, the parts of the code that draw the
actual content of the navbar have been separated out, and these
are simply not called when the navData list is empty.