[Bug 16844] Fix Mac OS X crash due to missing legacy theme font
In Mac OS X 10.11, the operating system uses the "San Francisco"
font its default UI theme.  However, this font is not licensed for
use in user documents -- only for use in user interface.
Requesting the font by name results in failure, and this failure
wasn't being detected, causing a null pointer deference when trying
to prepare fonts for the "legacy" theme.

This patch makes the legacy theme try to get a suitable font by
name first.  If that fails, it falls back to an 11pt system UI
font.