[[ Bug 16200 ]] Fix hang when choosing developer color from color picker on Mac.
The NSColor objects returned by the color picker on Mac might not
have a valid colorspace (e.g. the ones from the developer tab).
This patch fixes this by wrapping the colorSpace property fetch
in an exception handling block. If the fetch fails, the color
is converted to the NSCalibratedRGBColorSpace.