Set the LiveCode IDE to run in 32-bit mode by default
This can be changed by the end-user manually by toggling the
"Open in 32-bit Mode" option in the app bundle Info page. We still
need a way of re-launching the IDE in 32-bit or 64-bit mode though
as this is likely something that will need to change on a per-run
basis. (There are command-line utilities for running specific
architectures within app bundles).