Update MCField::GetFormattedHeight to be sensible
With the new text metrics, the formattedHeight value bore only a
passing relationship with reality. By tweaking the calculation
slightly (and, as a side-effect, better matching how heights for
field scrollbars are calculated), the formattedHeight now returns
a value that works for auto-sizing fields to match the height of
their contents.