MCFiber: Don't attempt to initialise/clear a pthread_t
The `pthread_t` type is opaque, and its structure is
implementation-defined.  The POSIX threads specification doesn't
provide any value for a `pthread_t` that is guaranteed to be an invalid
thread identifier.

Instead, modify the `MCFiber` to store the validity of the `thread`
field in a separate `have_thread` boolean flag.