[[ LicenseCheck ]] Add a simple license check API to externals V1 API.
The V1 externals API has been extended with a new API LicenseCheck().

This API takes the minimum edition (one of Community, Indy, Business)
which the external requires. If the engine is not licensed at that level
or above, it will cause the external handler to fail and raise an
error in the engine. If a call to the API fails in an external initialize
handler, then it will cause all future calls to the external to fail
with an error.

In addition to the new external API, the deploy command will now include
an edition byte in the standalone capsule - allowing license checks to
occur in standalones rather than just in the IDE.