[Bug 16744] com.livecode.canvas: Default to non-zero path fill rule
The default SVG path fill winding rule is non-zero, and the default
libgraphics path fill winding rule is non-zero, so make the canvas
library match!

Also, update the SVG icon widget to force the winding rule to
non-zero, and add a FIXME note that we should turn it into a property
in the future, probably.