[Bug 16145] Mark as a known issue.
On investigation with @livecodeian, it appears that the crashes are
due to an issue relating to a bug in the Android Gingerbread dynamic
linker with additional ELF sections.

Although a fix may be theoretically possible, we may not be able to
find one before LiveCode 8, especially since we do not have any way to
replace or modify the dynamic linkers on affected devices.  Ensure
that this is marked as a "known issue", in case it's not possible to
find a complete fix before 8.0 release.