[Bug 16927] com.livecode.browser: Don't show htmlText in prop inspector
Some webpages -- such as https://livecode.com/ -- have crazily
enormous inner HTML content.  The field takes a long time to lay it
out (even with wrapping disabled), which can make it look like the
engine has hung when opening a property inspector for a browser
widget.

It would be difficult to speed up the field adequately to address this
easy, but hiding the `htmlText` property seems like an adequate
workaround for now.