[Bug 17013] libfoundation: Don't crash constructing huge strings
Ensure that `__MCStringExpandAt()` checks for overflows and indicates
an out-of-memory error when an overflow occurs.  This prevents
horribleness occurring with the following script:

    put "0" into x
    repeat 32
       put x after x
    end repeat

Although the character count of a string is stored as a `uindex_t`,
this checks for overflowing an `index_t`.  This prevents two bugs:

* the construction of strings for which a negative index can't be
  represented

* the construction of a string `x` such that `len(x)` is negative

This fix exposes the fact that LiveCode 7 doesn't actually do anything
if an out-of-memory error is reported and blithely continues running
the script as if nothing has happened.