[[ LicenseCheck ]] Implement the ExternalV1 license check API (6.7)
This patch adds the LicenseCheck API to the 6.7 ExternalV1 interface.

As the current V1 glue code uses the interface version number to
decide if Unicode support is present we cannot bump that number in
6.7 - it must remain at 5.

So that the glue code can be updated to support 6.7, there is a new
context query for 'HasLicenseCheck' - if a call to query this does
not fail, then the caller can assume that the LicenseCheck API is
present in the function table.