widgetutils: Add getNativeThemeName() handler
Add a function that returns a theme name to be used by widgets when
the theme is "native".  This returns "iOS" on iOS and Mac OS X, and
"Android" otherwise.  For the time being, widgets may use this
function to determine how to automatically adapt themselves to the
current platform.