Set pseudo modal window layer to modal
NSColorPanel does not function correctly as a modal
dialog, however, we need it to behave as one.
Previously when `answer color` was called from a modal
dialog it would open in a layer behind and potentially
directly behind the calling dialog. This change
means the color dialog will open above the calling modal.
Unfortunately the color dialog itself opens modals so we
can't use a higher window layer to ensure it stays on top.