iconsvg: Divide built-in icons into families
Divide the icons into a "fontawesome" family and an "ide" family.
The "fontawesome" families icons are the only ones listed by
default, and therefore the only ones displayed in the icon picker.

This change is to avoid user applications becoming dependent on
IDE-specific icons, and will make it easier to place SVG path data
in resource files in the future.