[[ Bug 17090 ]] Ensure executing LCB code locks the host for execution.
If an object is executing a script then it cannot be deleted - this
change extends this logic to LCB code running as part of the object.

The 'scriptdepth' instance variable has been changed to be a 32-bit
integer (rather than a byte) thus meaning the MCfreescripts global
handling the overflow case can be removed.

Access to the scriptdepth is now via MCObject::lockforexecution()
and MCObject::unlockforexecution().

Calls to lock and unlock execution have been added to the places
where LCB code is called as a result of MCObject.