[[ ReplaceInField ]] Implement the 'replace in field' command.
This patch implements:

  replace ... with ... in <field chunk> (replacing | preserving) styles

It differs from the replace command in the following ways:

The <field chunk> is always evaluated as a field. This means that
'char 1 to 5 of tLongIdOfField' will be interpreted as char range
1 to 5 of the field which tLongIdOfField resolves to (rather than
of the variable).

If the preserving form is specified the replacement text takes the
style of the first char of the found pattern in each case.

If the replacing form is specified the replacement text takes no
style.