Don't attempt to use a CA file for libcurl outside Linux
The OSX server engine uses the system's copy of libcurl and that
uses the operating system's CA database rather than a CA bundle
as happens on Linux so trying to set a CA file path (even to the
default CA file path) causes curl to throw an error.

Windows libcurl requires a CA file to be supplied; it doesn't use
the system's certificate DB unless using Schannel (which ours
doesn't; it uses OpenSSL instead) and there is no default path to
a CA file bundle on Windows as the system doesn't supply one. As
such, we might as well remove the attempt to set it for Windows
too. (SSL doesn't work either way; we'll need some other
solution).