[[ Bug 16566 ]] Ensure smoothness of resizing on El Capitan
In order to prevent the SPOD in long running loops, the engine
'tickles' the event queue periodically (every 0.25s) to ensure
the OS does not think the engine has hung.

The tickling of the event queue can cause redraws to occur, which
is inappropriate whilst a resizeStack handler is running.

To fix this, the event checking is now turned off for the duration
of the handling of a resize event meaning that a redraw will only
occur at the end of its operation.