[[ Bug 17154 ]] Update mfocused when parent rect is changed
This fixes the issue seen in the tutorial where clicking on the
inspector menubar icon failed sometimes.

The reason is that before the mouse down message is sent, the
menubar performs a layout which temporarily moves controls in
a group underneath the mouse pointer (which updates the mouse
focus). The group itself is subsequently moved away, which does
*not* update the mouse focus, as MCObject::SetRectProp only
checks if the card focus is the object being moved.

This fixes the issue by recalculating the mouse focus if the moved
control is an ancestor of the focused control, rather than equal
to it.